requirements.txt
requests>=2.22.0
tox
pytest
python-dotenv
flit
black
